Removing isFile from speechres/commoncode
isFile is a commoncode function that Shanqing created to go along with
Audapter to be run with other versions of MATLAB. With the introduction
of MATLAB 2017b, isfile is a built-in file, and throws a warning now
that it finds two instances of the function in the path. In order to
correct for this, I am deleting it from the common files. This will
continue to work as long as you use MATLAB 2017b. The current
dependencies in the this toolbox of this repo are:
-commoncode/check_file
-commoncode/my_save
-audapter_matlab/mcode/dBSPL2WaveAmp
-audapter_matlab/mcode/read_parse_expt_config
-audapter_matlab/mcode/compareFormTrackParams
-audapter_matlab/mcode/UIRecorder
-audapter_matlab/mcode/runExperiment